+++ src/components/home/modals/CancelLease.vue
@@ -224,7 +224,7 @@ export default {
             }, response => {
                 this.errorMessage = response.body.message
                 if (this.errorMessage === undefined) {
-                    this.errorMessage = 'Unknown.Please check network connection!'
+                    this.errorMessage = 'Unknown. Please check network connection!'
                 }
                 this.sendError = true
             })
